requirements.txt
Flask==2.3.3
MarkupSafe==2.1.5
python-docx==0.8.11
openpyxl==3.1.2
pypdf==5.0.1
pdfplumber==0.11.7
pdfminer.six==20250506
PyMuPDF==1.24.0
Werkzeug==2.3.7
chardet==5.2.0
pytesseract==0.3.10
Pillow==10.4.0
pdf2image==1.17.0
rarfile==4.2
pytest-timeout==2.1.0
requests==2.31.0
xlrd==2.0.1

# RAG и векторный поиск
psycopg2-binary==2.9.9
pgvector==0.2.5
openai==1.54.0
httpx==0.27.2
tiktoken==0.8.0

# ORM, миграции и безопасность (инкремент 13)
SQLAlchemy==2.0.35
alembic==1.13.3
cryptography==43.0.3
bcrypt==4.2.0
PyJWT==2.9.0

# Планировщик задач (инкремент 15)
APScheduler==3.10.4
python-dotenv==1.0.1

# Рендеринг Markdown и санитизация HTML
bleach==6.2.0
markdown==3.7

# Опциональные зависимости для модуля анализа текста
# spacy>=3.0.0
# ru_core_news_sm (устанавливается через: python -m spacy download ru_core_news_sm)